+++ packages/fieldsConfig.js
@@ -298,24 +298,10 @@ export default [
   {
     title: '插件字段（富文本需要单独引入）',
     list: [{
-      title: '分割线',
-      type: 'divider',
-      labelWidth: '40',
-      prop: 'divider',
-      component: 'elDivider',//ele分割线
-      span: 24,
-      icon: 'icon-map',
-      params: {
-        html: '<h3 style="color:red">分割线标题</h3>',
-        contentPosition: "left",
-      }
-    }, {
       title: '日历',
-      type: 'calendar',
-      labelWidth: '40',
-      prop: 'calendar',
-      icon: 'icon-map',
       component: 'elCalendar',//ele日期
+      labelWidth: '40',
+      icon: 'icon-map',
       span: 24,
       params: {